requirements.txt
# Core Framework
fastapi==0.115.6
uvicorn[standard]==0.32.1

# Database ORM
sqlmodel==0.0.22
psycopg[binary]==3.2.3

# Authentication
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0

# Validation
pydantic[email]==2.10.3

# Environment
python-dotenv==1.0.1

# AI Agent Framework (Phase III) - Using Google Gemini
google-generativeai>=0.8.0

# Development Dependencies
pytest==8.3.4
pytest-cov==6.0.0
pytest-asyncio==0.24.0
httpx==0.28.1
